+++ src/icon/comfortSeat.tsx
@@ -1,9 +1,9 @@
 // tslint:disable:max-line-length
 import React from 'react'
 
-import StatusIcon from '../_utils/icon/status'
+import StatusIcon, { IconPropsWithStatus } from '../_utils/icon/status'
 
-export const ComfortSeat = (props: Icon) => (
+export const ComfortSeat = (props: IconPropsWithStatus) => (
   <StatusIcon {...props}>
     <g fill={props.iconColor} fillRule="nonzero">
       <path d="M4.41 4.063l-1.423.342a1.295 1.295 0 00-.96 1.525l.07.17 8.49 13.14.007.017.09.132c.255.361.642.6 1.055.634l7.976.004A1.29 1.29 0 0021 18.732v-3.14l-.007-.132a1.29 1.29 0 00-1.278-1.162h-6.456l-.686-1.45c-1.012-2.106-1.374-2.662-2.462-4.117-1.182-1.58-2.363-2.764-4.105-4.223-.429-.36-1.095-.566-1.596-.445zm7.084 14.8l-.067-.165L3 5.656l.008-.062a.29.29 0 01.213-.217l1.423-.342c.164-.04.518.07.72.24 1.558 1.304 2.64 2.369 3.685 3.713l.26.342c1.13 1.51 1.433 1.986 2.614 4.48l.7 1.488h7.092a.29.29 0 01.285.295v3.14a.29.29 0 01-.285.294l-7.822.001a.465.465 0 01-.391-.215l-.048-.07-.014-.024.054.143z" />