+++ src/icon/eyeIcon.tsx
@@ -8,7 +8,7 @@ export interface EyeProps extends Icon {
   readonly off?: boolean
 }
 
-export const EyeIcon = (props: EyeProps) => (
+export const EyeIcon = ({ off, ...props }: EyeProps) => (
   <BaseIcon {...props}>
     <g
       fill="none"
@@ -17,13 +17,13 @@ export const EyeIcon = (props: EyeProps) => (
       strokeLinecap="round"
       strokeLinejoin="round"
     >
-      {props.off && (
+      {off && (
         <g>
           <path d="M17.94 17.94A10.07 10.07 0 0 1 12 20c-7 0-11-8-11-8a18.45 18.45 0 0 1 5.06-5.94M9.9 4.24A9.12 9.12 0 0 1 12 4c7 0 11 8 11 8a18.5 18.5 0 0 1-2.16 3.19m-6.72-1.07a3 3 0 1 1-4.24-4.24" />
           <line x1="1" y1="1" x2="23" y2="23" />
         </g>
       )}
-      {!props.off && (
+      {!off && (
         <g>
           <path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z" />
           <circle cx="12" cy="12" r="3" />