+++ src/icon/pinIcon.tsx
@@ -9,16 +9,11 @@ export interface PinProps extends Icon {
   readonly strokeColor?: string
 }
 
-export const PinIcon = (props: PinProps) => (
+export const PinIcon = ({ bgColor, strokeColor, ...props }: PinProps) => (
   <BaseIcon {...props}>
-    <g
-      stroke={props.strokeColor}
-      strokeLinecap="round"
-      strokeLinejoin="round"
-      strokeMiterlimit="10"
-    >
-      <path fill={props.bgColor} d="M20 9c0 4.9-8 13-8 13S4 13.9 4 9c0-5.1 4.1-8 8-8s8 2.9 8 8z" />
-      <circle fill={props.strokeColor} cx="12" cy="9" r="3" />
+    <g stroke={strokeColor} strokeLinecap="round" strokeLinejoin="round" strokeMiterlimit="10">
+      <path fill={bgColor} d="M20 9c0 4.9-8 13-8 13S4 13.9 4 9c0-5.1 4.1-8 8-8s8 2.9 8 8z" />
+      <circle fill={strokeColor} cx="12" cy="9" r="3" />
     </g>
   </BaseIcon>
 )